chore(interpolateFileName): extract to dedicated file
